Added rawTypeString function to help debugging
The typeString function requires knowledge of Delta so that it
can find the names of type aliases. For debugging, however, this
is annoying since it calls solve and also because we might not
have access to Delta. So, this function prints the raw type strings
without aliases or pattern recognition.