Added basic unit testing system
Unit tests can be placed in the new tests directory and annotated
with a description and expected output. The check.rb script will
run every .lj file in the tests directory with the interpreter
and compare the results to see if the test passes.

A new "check" recipe has been added to the makefile that automates
running the check script.