Added RECVAR type to make MU typings more reliable
Now, instead of overloading TYVAR, the variable in each MU type
is represented by its own type, RECVAR, which contains a reference
to the parent MU type.

This allows the type system to better differentiate between actual
TYVARS and those that are only meant to act as markers for MU types.