Fixed a bug where recursive definitions would sometimes not check
If a val definition contained a self-reference but did not start
with a lambda expression, it would not type check because the
type of itself was not introduced into the environment.

Now, the type of a definition's own name is included in the typing
environment independent of if it is a lambda expression.