Added DEBUG structure to AST file
The DEBUG structure contains the typeString function which
simply prints a simple representation of a type tree. Since
the typeString function inside lc.sml is so complex, its
nice to have a simple version that doesn't have so many
dependencies.

In the future other debugging functions with access to AST
may be placed here.